+++ engine/common/recaptcha/__init__.py
@@ -1,0 +1,1 @@
+from .recaptcha_v3 import check_recaptcha_internal_api  # noqa
