+++ engine/common/tests/test_recaptcha.py
@@ -2,7 +2,7 @@ from unittest.mock import patch
 
 from django.test import override_settings
 
-from common.recaptcha import check_recaptcha_v3
+from common.recaptcha.recaptcha_v3 import check_recaptcha_v3
 
 action = "test_action"
 score = 0.9
@@ -13,7 +13,7 @@ client_ip = "192.168.1.1"
 @override_settings(RECAPTCHA_V3_ENABLED=True)
 def test_captcha_v3_fails():
     with patch(
-        "common.recaptcha._submit_recaptcha_v3",
+        "common.recaptcha.recaptcha_v3._submit_recaptcha_v3",
         return_value={"success": False, "error-codes": ["invalid-input-secret"]},
     ):
         valid = check_recaptcha_v3(value, action, score, client_ip)
@@ -24,7 +24,7 @@ def test_captcha_v3_fails():
 @override_settings(RECAPTCHA_V3_ENABLED=True)
 def test_captcha_v3_fails_actions_dont_match():
     with patch(
-        "common.recaptcha._submit_recaptcha_v3",
+        "common.recaptcha.recaptcha_v3._submit_recaptcha_v3",
         return_value={
             "success": True,
             "challenge_ts": "2023-03-02T14:44:18Z",
@@ -41,7 +41,7 @@ def test_captcha_v3_fails_actions_dont_match():
 @override_settings(RECAPTCHA_V3_ENABLED=True)
 def test_captcha_v3_fails_score_too_low():
     with patch(
-        "common.recaptcha._submit_recaptcha_v3",
+        "common.recaptcha.recaptcha_v3._submit_recaptcha_v3",
         return_value={
             "success": True,
             "challenge_ts": "2023-03-02T14:44:18Z",
@@ -60,7 +60,7 @@ def test_captcha_v3_fails_score_too_low():
 def test_captcha_v3_fails_invalid_hostname():
     hostname = "localhost"
     with patch(
-        "common.recaptcha._submit_recaptcha_v3",
+        "common.recaptcha.recaptcha_v3._submit_recaptcha_v3",
         return_value={
             "success": True,
             "challenge_ts": "2023-03-02T14:44:18Z",
@@ -77,7 +77,7 @@ def test_captcha_v3_fails_invalid_hostname():
 @override_settings(RECAPTCHA_V3_ENABLED=True)
 def test_captcha_v3_valid():
     with patch(
-        "common.recaptcha._submit_recaptcha_v3",
+        "common.recaptcha.recaptcha_v3._submit_recaptcha_v3",
         return_value={
             "success": True,
             "challenge_ts": "2023-03-02T14:44:18Z",
@@ -95,7 +95,7 @@ def test_captcha_v3_valid():
 def test_captcha_v3_valid_with_hostname_validation():
     hostname = "localhost"
     with patch(
-        "common.recaptcha._submit_recaptcha_v3",
+        "common.recaptcha.recaptcha_v3._submit_recaptcha_v3",
         return_value={
             "success": True,
             "challenge_ts": "2023-03-02T14:44:18Z",
